Implemented basic callback support for subscribing to event based callbacks.
This method currently lets you subscribe to create/update events for all
models and specify an exec that should take place, as well as data that should
be passed to the exec command. It also allows filtering based on the results
of boolean methods. The two test variations are notify me when my box finishes
provisioning, and notify me when an emergency log message comes in.