First pass at IPMI power management.
I stupidly used the SoftLayer API documentation when developing the initial
power management plugin which had notions of powerCycle and rebootSoft, where
power cycle was documented as being like pulling the power and plugging it
back in. IPMI doesn't have this notion. rebootSoft really needs to be
gracefulShutdown and powerCycle (which is graceful, it uses ACPI) should be
updated in the UI to describe what it actually does.