Moved to akka 2.0.1
What took so long to debug is that the underlying configuration system now
supports lists. Previously our config wrapper could call getString on call
keys and get back a string value. Now, if you do that and the underlying value
is a list an exception is thrown. Put in a hack to skip akka configs but will
need to actually fix this. Should also move to using lists instead of our hack
of having comma separated values in a string.