Fixed power component ordering
When you use a list comprehension on the power units/components in a view, the
result ends up being reordered due to the conversion of the component into a
form element. After moving the creating of the components set into a helper, I
was only seeing one component per power unit. This was due to the fact that
the compare method and the equals method were inconsistent. Compare was used
for ordering, and thep revious compare method would view two components in
the same power unit as equal for sorting purposes. Introducing a position
method for the power component (as in, the position of the component with
respect to the unit) fixed that issue.

Fucking equals/compare/hashCode strikes again!