Allow callback commands to be specified as lists or strings
Scala sys.Process uses ProcessBuilder under the hood which just breaks up
commands and args by whitespace. If the command isn't something like 'cmd
--arg1=val1 --arg2=val2' where splitting on whitespace results in a valid
command, such as for printf (which may have a whitespace separated format
specifier) it is useful to allow the command to be explicitly specified as a
list.