Make callback handling non-blocking
This was actually causing a somewhat interesting (and dumb) problem where we
would start a SQL transaction (to for instance update an asset), then the
callback would be triggered. If the action for that callback looked up the asset
in collins via the API, the transaction would not yet have committed and the
changes would not yet be visible. This change moves the callback processing to
be handled asynchronously so that the executed callback action should occur
after the transaction is committed. Duh.