Easier keyboard navigation
Set the tabindex so that after the asset tag, the column on the right is tabbed
through before the column on the left.

There is an accesskey for each column. l (left) for the left hand column, r
(right) for the column on the right. On the mac, using Chrome, ctrl+alt+l and
ctrl+alt+r will swap between columns.