Parse server and store base server information
For lack of a better term, we will call the "thing that holds a
motherboard" the base server and keep track of the basic asset
information (description, product, vendor).  Description is less
likely to be useful, but it is easier to include it than not.  Unlike
cpus/nic/disks there is only one bit of basic server information (aka
it's not a Seq).

(Note that all dependencies do not yet compile, but this is the most
self contained unit for just the parsing code.)