Remove the use of arbitrary dispatcher and use play built in actor system
Summary:
After our fumbles with ipmi errors, reviewed the implementation and use
of dispatcher. This refactoring makes proper use of plays actor system.
We can document the actors when they are used. The actors can be configured
using plays configuration (same as the previous attempt). Collins
now only requires the default-dispatcher on account of using
`Execution.Implicits` at a few places. The remainder of Collins relies on
actors