Reorganize structure of Collins
Summary:
After having upgraded play a few time, I have come to the
realization that the organization of the project may be an
hinderance. An ide would make the upgrades easier to
achieve.

This is a first step of upgrading to play 2.2.6 and
was done as a series of 36 commits.

One big advantage for me, is that I can import the project
into eclipse without any errors (red markers) and may be
in the future make refactoring the crufty parts easier.