Address issues with Mixed Authentication provider.
Summary:
When we merged in the changes for MixedAuthenticationProvider, we
did not realize a couple of issues.

MAP accepts an authType that is ',' seperated. Which allows
for "file,ldap". However the config validation was not updated to
account for this.

This manifested in the unit tests with mixed mode test "file,ldap"
where ldap configuration was not provided.

This commit addresses these issues, when specifying "file,ldap"
authentication mode, both file and ldap configuration MUST be
valid.