Support tattler thru the secure controller
Summary:
Currently tattler log source selection is arbitrary, i.e. depends
on the type of tattler in use. An api call can use user source or
api source for tattling. Provide a standard means that ensures
correctness of tattler source.

In addition, capture the user causing the tattle in a dedicated
field on asset log instead of as part of the message.