Introducing hazel cast for clustered operation of collins
Summary:
In standalone (1 instance) mode, collins relies on GuavaCache for
in-memory cache. This prevent operation in a clustered environment
where it is not possible to ensure cache is coherent across multiple
instances, which can only be achieved by reducing ttl and therefore
the effectiveness of the cache.

As part of doing this, fixed incorrect assumption and implementation
to allow for the domain objects to be appropriately serialized.

Hazelcast will serve as the pub/sub module for firehose implementation.

Note:
Hazelcast was introduced to support distributed message queue for firehose
The Firehose relies on collins callback mechanism, this commit includes
changes to fix up the callbacks.

2 specifications will many tests have been added as part of this effort
for cache and callbacks.