Load runners lazily
That is, wait until we are going to use a runner before we load it.

We have seen issues where running consolr a lot on a server will make the server run out of resourers.

We think loading runners with lots of dependencies takes a lot of resources, so I want to try to load
the runners we use more often earlier, and not load bigger, less used runners if they aren't needed.