Add checkbox to redirect New assets to intake
This adds a "Intake Redirect" checkbox on the index page. When checked
and the search returns a single asset with status New, the browser get
redirected to the intake process.

This replaces the old behavior where a search for only the tag always
redirects to intake process if the asset is New.